+++ reactproject/src/components/AppRouter.jsx
@@ -2,12 +2,14 @@ import React, { useContext } from "react";
 import { Routes, Route, Navigate } from "react-router-dom";
 import Main from "../pages/Main"
 import Auth from "../pages/Auth"
+import Settings from "../pages/Settings";
 import { Context } from "../index";
 import { useAuthState } from 'react-firebase-hooks/auth'
 
 export const routes = {
     AUTH: '/',
     MAIN: '/main',
+    SETTINGS: '/settings'
 }
 
 const AppRouter = () => {
@@ -20,6 +22,7 @@ const AppRouter = () => {
                 user ?
                     <>
                         <Route path={routes.MAIN} element={<Main />} />
+                        <Route path={routes.SETTINGS} element={<Settings />} />
                         <Route path="*" element={<Navigate to={routes.MAIN} />} />
                     </>
                     :
