+++ reactproject/src/pages/Settings.jsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import SettingsCard from "../components/settings/SettingsCard";
+
+const Settings = () => {
+
+
+    return (
+        <SettingsCard />
+    )
+}
+
+export default Settings;
